// backend/src/models/enums.ts
export enum FunctionType {
  BESSEL_FIRST_KIND = 'bessel1',
  GAMMA = 'gamma',
  BETA = 'beta',
  LAGUERRE_POLYNOMIAL = 'laguerre',
  LEGENDRE_POLYNOMIAL = 'legendre',
  JACOBI_POLYNOMIAL = 'jacobi',
  CHEBYSHEV_FIRST_KIND = 'chebyshev1',
  CHEBYSHEV_SECOND_KIND = 'chebyshev2',
  HERMITE_PHYSICIST = 'hermite1',
  HERMITE_PROBABILISTIC = 'hermite2',
}